+++ acme.py
@@ -51,6 +51,7 @@ def parseArgs() -> argparse.Namespace:
 	parser.add_argument('--import-directory', action='store', dest='importdirectory', default=None, metavar='<directory>', help='specify the import directory')
 	parser.add_argument('--network-interface', action='store', dest='listenif', metavar='<ip address>', default=None, help='specify the network interface/IP address to bind to')
 	parser.add_argument('--log-level', action='store', dest='loglevel', default=None, choices=[ 'info', 'error', 'warn', 'debug', 'off'], type=str.lower, help='set the log level, or turn logging off')
+	parser.add_argument('--headless', action='store_true', dest='headless', default=None, help='operate the CSE in headless mode')
 	
 	return parser.parse_args()